fix bitlist encoding; rm indirect marshal dispatch
MarshalSSZTo was being called through ssz.MarshalSSZ. This change
inlines the work MarshalSSZ is doing. Note that this was causing
false confidence in spectests, because new code was unmarshaling
but not marshaling in the round trip test.
Fixing the spectest turned up a bug in bitlist marshaling, which
has been fixed.